requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.0.350
langchain-community==0.0.38
langchain-groq==0.1.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pypdf==3.17.1
python-dotenv==1.0.0
httpx==0.25.2
jinja2==3.1.2
websockets==12.0
structlog==23.2.0
africastalking==2.0.1
requests==2.31.0